Pivot tables on subscriptions.
The pivot tables show us the summaries of the subscriptions.

The first table show us the summary of the subscriptions per subscription type.  The subscription type being basic, premium, and standard.
The second table shows us the summary of the subscription per year in the different subscription type.
the third table shows us the summary of the subscription per region, and it gives us insight on which subscription type each region prefers or uses most.